+++ src/common/base-chart.component.ts
@@ -57,6 +57,8 @@ export class BaseChartComponent implements OnChanges, AfterViewInit, OnDestroy {
   update(): void {
     if (this.results) {
       this.results = this.cloneData(this.results);
+    } else {
+      this.results =  [];
     }
 
     if (this.view) {